requirements.txt
asgiref==3.8.1
certifi==2025.10.5
cffi>=1.17.0
charset-normalizer==3.4.3
cryptography>=42.0.5
defusedxml==0.7.1
diff-match-patch==20230430

Django==5.0.3
django-cors-headers==4.3.1
django-datatables-view==1.20.0
django-extensions==3.2.3
django-filter==24.3
django-import-export==3.3.7
django-jazzmin==2.6.1
django-rest-knox==4.2.0
django-serverside-datatable==2.1.0
django-simple-history==3.5.0

djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.10
MarkupPy==1.14

numpy
pandas

odfpy==1.4.1
openpyxl==3.1.2

packaging==24.0
pillow>=11.0.0

psycopg2-binary==2.9.11
pycparser==2.22
PyJWT==2.8.0

python-dateutil==2.9.0.post0
python-decouple==3.8

PyYAML==6.0.1
requests==2.32.5

six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2024.1
urllib3==2.5.0

xlrd==2.0.1
xlwt==1.3.0